+++ lib/src/elements/sink/renderer/windows/wasapi_renderer.rs
@@ -633,7 +633,7 @@ impl Sink for WasapiRenderer {
                     master.reset_for_seek().map_err(WasapiRendererError::from)?;
                 }
             }
-            ControlMsg::Seek(_) | ControlMsg::CheckSeek(_) | ControlMsg::Preroll(_) => {}
+            ControlMsg::Seek(_) | ControlMsg::Preroll(_) => {}
         }
         Ok(())
     }