+++ lib/src/elements/source/compositor/sw_video_compositor.rs
@@ -513,7 +513,6 @@ impl Sink for SwVideoCompositorInputSink {
             ControlMsg::Pause
             | ControlMsg::Resume
             | ControlMsg::Seek(_)
-            | ControlMsg::CheckSeek(_)
             | ControlMsg::Preroll(_) => {}
         }
         Ok(())